requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
